ping-monitor can use multiple AMQP hosts
Update the ping monitor to connect to AMQP using the Baton::Channel
class, this ensures that it will get the support for connecting to
a pool of RabbitMQ servers.

Defined a new input exchange as Baton::Channel needs one and I thought
it was best to be seperate from the baton_ping_in exchange.